Use fixtures defined in a central place
